Assignment2/requirements.txt
gensim
sklearn
torch
pickle
transformers
sentencepiece
nltk
indic-nlp-library
tarfile
